Add upto 69 features to the training pipeline
Unfortunately, the results did not improve. What does that mean?

It means, first, that the metric we want to predict (aka price in 5 minutes)
is very hard to predict. Some people say that short-term prices are completely random.
I think there is a chance to catch signal and build a model, however
with the current set of featues derived from OHLC candles we are not
moving forward.

What can you do to fix it?
> Pull more data sources For exaample, it can happen that  several crypto
currencies tend to co_move. So, generate OHLC candles for other crypto products
and add them as featues to the training pipeline.

> Instead of predicting prices directly, you try to prediict spreads between
prices, Something called as "pair trading"