Add model_version to PricePrediction
We also tried to add the git commit hash to the Docker image, but didnt work

Things work when you run the APi without Docker, because the "git rev-parse HEAD"
command returns the correct git commit hash.

When we run with Docker we want to pass hash using environent variable
durin the build process. Things look right but dont work